zs_completion: make '\t' (tab) consistent with the other types of completions, to avoid the slight confusion and second thoughts when performing particularly a filename completion.
This has been proved after practice and it was annoying. Now ' ' (space) does the selection
and '\t' (tab) acts as an arrow down key. When the selected item is a directory though, the
autocompletion continues automatically. When this is not desirable '\e' (escape) returns to
the previous state (a very small price to pay, plus is consistent with the kind of completion
that is performed to our editor, so the brain has been trained to those expectations).

This was the last remaining issue, as a decision had to be made. And came quite naturally.