Add a memory interface to libc.
This pulled from:
https://github.com/miguelperes/custom-malloc
at :commit a07735c5ef3f467b0b4f1cf0fb217e737e74a3a7

It should be noted that initially has been tried the code from:
https://github.com/RAGUL1902/Dynamic-Memory-Allocation-in-C
with the exact logic. This is condireded as an important factor
to the realization off the mechanics.

Equally thank both and anyone involved.

Note that the original code has been adapted to this environment.

Add also a memtest-module.

It is too early to judge, but this looks sufficient and functional.