Add an initial FILE (file pointer) implementation, that is a higher interface than file descriptors, alas with an overhead.
We've tried to minimize this overhead a bit, though it is really an introduction.

And quite probably the implementation might misbehave at points and with variation
types of data arguments. The worst case missing of information. This needs time
to think for things to check for a quite broad variable set of conditions, as it
is easy for many things to go wrong with I/O. Quite probably the code will do it
right, but an unknown external factor can influence the result. So because it is
the heaviest responsibility to do the communication stuff (the outside and inside
type of things) for your ecosystem, this will get time. But we do not really have
so much time to wait, and so we had to do something.

And merely because it is very convenient with streams and many packages they need
the functionality, and so delayed the migration to libc.

This was possible thanks to the implementation from neatlibc and a little bit help
from the friends of minibase and dietlibc.

We've also had to modify the vsnprintf() implementation to accept a callback function
to send its byte. By default it writes the byte to its buffer first argument.

fprintf() though, that requires to parse directives and a variable set of arguments,
and prefers to use vsnprintf(), it sets its own callback that does its thing.

This function is to decide to either store this byte in a temp buffer (instead of
calling sys_write() repeatedly), or call fflush() (if the buffer is full, or the
character is a new line character, or the buffer type is in an unbuffered state
(like the standard error stream is by default)). This means that the standard error
stream it writes imediatelly the byte to its underlying file descriptor.

And does its stuff, by working with its filePointer (FILE) structure. It gets its
pointer, by accessing and casting the userdata (void *) field, from the low level format function
(that actually does the parsing) first argument.

This function has been exposed as str_format() and its structure as fmtType.

We really like this vsnprintf() because of its really hard work to deal with all
this complexity, and we would like to keep it forever. I took the currage and the
irresponsibility to migrate some of the code to our style. Really sorry about that,
but you may like to know that i only did it, because it helps me concentrate when
dealing with an application myriad types of challenges. And a same usual symbol
names and a routinally (almost) same set of actions and logic flow helps a lot of
bits.

So there is nothing wrong with your style (in any case its only a style, what it
matters at the end, its the code, and this specific function obviously got a lot
of work and care, and I feel grateful).

Thank you all anyway.

This commit is incomplete but it is late night, and i've to wake up early, and i
really got really tired - but happy.

Rrrreally it needs a lot of work yet, but it is functional!

Have a good now!