scdoc: distinguish document errors than EOF and return proper value to the caller.
Rename ustring to utf8.

Generate.libc: allow .c units, to be treated as metapackages (do not throw error,
in the case of empty body and or without exposed function in the unit header).

But:
We're facing a crash in logic when compiling modules that link against libraries
that uses the memory interface from own libc. Specifically when the library allocates
an object (like string *) and returns it to the module and the module reuses this
object, the language (that has been compiled against the standard memory interface
from standard libc) it will send this chunk of memory to standard free (), that it
doesn't know how to handle it, as there is no compatibility.