scdoc-module: do not use libc functions and duplicate the output string in scdoc_parse().
This is to avoid the crash that has been mentioned in the previous commit:
3dea77268cfff7e8babd3610688ffd259930b448

Use the stdio libc interface in independent libraries and modules.

Make: enable verbosity by default so replace the option with --no-verbose.

Expose the filetype-module as File, and handle this situation also to the
standard file-module from the z namespace.

Add a file_read() function to the file-module that returns a string.

Realize that string_new_with_allocated() from libstring, is a bit weak interface,
as it assumes that memory size is len + 1.

It is being used three times in the distribution, usually from modules to
send back a string. As long it is being used from modules it's rather ok,
but in libc we now require an extra size_t argument to buffer size.

Document all the new modules that has been produced during development
outside the z namespace. Those all require the Make utility to build.