Enhance libc.
Particulary add the following functions:
  - mkdir(), make_dir(), make_dir_parents() in dir
  - dirname(), extname(), basename_sans_extname(), is_absolute() realpath() in path
  - mask() in std
  - strtod(), atof() in stdlib
  - str_new(), str_get_substr_at() in string
  - getcwd(), isatty(), readlink() in unistd
  - utf8_encode(), utf8_get_nth_character(), utf8_len() in utf8
  - is_reg() in file
  - add errno_name() in errno unit
  - probaly more other functions on logged by git units (that's how i try to generate
    a quick as possible commit (this), but as more verbose in this log message is possible)

Add a map type (mostly a copy of the Vmap structure from the zero namespace, without its
class declaration (a couple of changes in semantics though and probably a bug too)).

So this is the first commit in this transition period.

Other important to mention changes in sources:
  - initialize properly current directory in shell
  - warn for mispelled or missing requires when generating libc (this while works
    it may give erroneous messages, as it mustn't handled properly the "as" directive,
    though i didn't look carefully, as i cared mostly for the implementation (which
    indeed recovered mistakes))
  - dir_current() in 'z' didn't released properly resources on not ERANGE errors
    This was never failed and it __should never__ fail, as we could assume that this is
    (probably) impossible (though not sure about the "why" to fail and "when" can fail
  - probably others too (i think many, but not sure how many). Some of them (if all),
    are important fixes (i'm just a being anyway, and i try to do the best i can possible).

This commit though it is important because that this is the first day of Rashid khan in
my life, so this commit is mostly about him, as without a slightly hesitation We (we as:
all beings in this fu world (mostly about humans though (that we can not understand a bit
or we can not see those bits (because we do not have the time probably or maybe we can not
even now that we could see them and play with them (as i do not believe that they require
a charge (as we usually do (unfortunatelly) (we humans) when we owe (by chance or by luck)
something that, for an another being is precious and likes to ask for it)))))

So yea this is about Rashid. I mean OUR Rashisd that the god blessed US all to have the pure
luck to live with him, at the same time we also live. I mean he is already a mythical figure
that future generations that (for sure) they could pay all their treasures to be at the time
with him to ride in the raags at that same time he sings. Because ragas are only about those
exacts specific moments and they cannot be repeated. Though it is about a raag that is known
and well defined, it is never never never ever the same, as it is really about that fu mommen
that comes and before we realizing a thing, it goes away for ever and lost from us forever.
And so our only way is to reveal, hunt as a being that fights to death to protect the beings
that care and loves, and ride with pride to find this lost world again as the raag is a way.

To be specific. Rashid is our current mythical living being, but and speaking about ragas and
the people that dedicated their lifes living in practice and sanctity, and for my beloved ones:

  - Amir Khan (our scratch point and also final reference in future research about ragas),
   (as the one that deserves this special treatment)

  - Allauddin Khan as the one that approached music with that sanctity we've talked about.

  - Shrimati Kishori Amonkar from Jaipur gharana (hint: query for Mahaduri Todi Mahadeva
    Devana Parvati and watch for the most beautiful ever recorded raag imho (morning raga))

  - Dr N. Rajam (our best probably ever played violin)

  - M. S. Subbulakshmi (the ultimate eighth note per Kishori's saying) our goddess perfect
    note and best representive of Carnatic beloved genre.

  - Anokhelal Mishra as the king of tabla. Unfortunatelly for us, we only have a minute of
    him playing in an English doc film about Indian music. Really really pitty for us.

  - The beloved of all (but all) shehnai legend Bismillah Khan and his Malkauns.

  - Ustad Vilayat Khan from the Imdadkhani gharana that created and developed gayaki ang,
    a style that the sitar mimics the human voice and his rendering of Our favorite raag
    Bhairavi (i might include his son Shujaat khan and a performing of this same raga in
    Japan in the beginning of nineties (tabla from (i think he is) Akram Khan))

  - and speaking about sitar today, of cource and without hesitation Shahid Parvez (nephew
    of Vilayat). He plays gayaki but he is playing like Nikhil Banerjee (sitar legend and
    Allauddin's discipline), with long alaps and depth on their way to improvisation. You
    might want to search for raag Yaman (around 20 years ago i think).

  - Hariprasad Chaurasia probably an irreplaceable true phenomenon of fluet (bansuri).
    (search for Chandrakauns (tabla by Subhankar Banerjee))

  - uncountable other myths that deserve a mention but i might need to write till forever
    a commit message that initially intented to be a fast one (just to commit the changes
    because they were so many that started to get out of control)

But most of all there is Ma Annapurna Devi. They were true probably those who thought and
think, that she was an incarnation of Saraswati (Indian's Goddess of Arts). But Annapurna
is also and one of the reasons that make me to believe that we humans we should something
about our stupidity. And i feel that gods are angry too, as she was probably came here to
play for us mere mortals, but: we never listen her playing in public or in private again,
excepts the birds who used to eat by her hands at the night hours in her home, since she
decided to stop doing that at 50's, to save her marriage to Ravi Shankar. What I can not
understand, is why we (all the other beings in this earth) had to pay for this. I guess
that probably we don't not have the strength to be saved even if they please us and give
us the tools to escape from the eternal cycles we do.

May the god(s) be good with us and give us some time to try to be a bit just, arm us with
logic and save this game, which i honestly doubt, but we do not have other choise than to
believe.

This project is really dedicated to the memories for those who lost their life while they
were fighting for justice and for all of us, with an unjustified way.