When generating libc, double the #ifdef headers, so the compiler will parse them at least twice.
This is to avoid re-require functions from the compilation units, when a required
function is too far at the end, and that function depends on other preceded functions.
I think to be sure we need at least three passes but probably we are fine with two.
In any case as a last resort, we can 'require' them from the compilation unit, as
the compiler will complain anyway.

Also add guards to almost all defined macros and interfaces (by default are defined),
so they need a "WITHOUT" definition, to be excluded from the generated source.

Add more functions to libc.

Particularly:

 - str_cmp()
 - str_cspn()
 - fstat()
 - getgid()
 - getuid()
 - getegid()
 - geteuid()
 - std/int.h (stdint.h)