This is the first major development to the memory interface, towards stabilization.
It fixes many bugs to almost all of its functions.

Now by default any allocated memory chunk is filled with zeros before returning to
the caller.

We still do not require a prior initialization before we use the interface, but
this might change in the future.

The calloc() function has been removed as it is never being used in our code.

The check for integer overflows has been removed, because its code was week as it was
involved an unsigned division. Looking around seems gnulib does it right, but the
code is rather complicated for our tastes, but it tries to handle all the possible
conditions that can be met, but with a price, albeit small. Maybe we'll end up to do
a really simple check, rather than to resort to complication (as we don't have the
qualifiers to understand the glory details from the C experts). Anyway our requirenments
in our codebase, are not for such huge memory quantities and we don't have to deal
with calloc() anymore. In any case we do what calloc() does.

During the development revealed bugs from applications that were hiding before,
and so it was really hard to debug and it took time to distinguish the actual bugs
coming from this interface.

Note that this is really the first revision, and as such it is considering as an
introduction, but probably and because of its pure simplicity, if it works we don't
have much to do, if at all. Probably future development will abstract the low level
code with a higher interface that might envolve separate pools of memory that don't
have to interfere with its other, so a structure might be required that will handle
those details that will provide this flexibility. Time will tell.