Enhance libc.
Particularly add:
  - tcgetattr()
  - tcsetattr()

and

the terminal interface, an exact copy of the TermType from the z namespace,
migrated to our libc. This is the first i think library that is independent
from standard libc.

But this seems to be a special commit, as with this and the recent additions,
we've achieved a succesfull build of our language with the home made libc.

May you have an eternal good now.