This is an upolished first language code that succesfully builds with own libc.
However, though it is the same (almost) code, modules should be included at the
building time, as we've failed hard with the dynamic load code that we've tried
hard, so there is no dlopen() and friends and i'm afraid that i don't know what
and how to do it. Maybe if time allows and refound my lost courrage, to go back
and steel hard an exact copy of musl or even glibc, otherwise i don't know.

Commited also an exact copy of the old tests without the import tests, that it
passes them all, except the known failure on the double type operation.

A fisrt benchmark with those two versions, shows a serious decrease in terms of
performance, when running those tests with the executable that is using our own
libc (around 7++ times slower, even with a optimization in the memory interface
(when we preallocate space at the initialization), but this was expected though
it could have been better, as there is no way to compete with the overoptimized
code written, during the last 30++ years of the glibc evolution, by C experts).

Nonetheless seems that is a kind of a special point, as we walk the development
path, as we can do now to use from point zero++ (compiler & linker) inner tools
to produce a very tiny system that can be expanded quickly to do at least basic
machine administration.

But also, and in my humble opinion, it is the first language that i know of that
is independent. And without to be that... poor, on the contrary i believe. It is
both expressive with nice and simple imho syntax, with a rich set of datatypes.

Anyway what we want from it, it is to be the glue and drive C to bind an endless
in quantity and quality C universe (even if we could freeze the time with a way,
there is unimaginable material anyway, and of course is and our product, that it
can be the source and a referenche itself - if it was in eighties i bet it could.