shell: '?' on completion menu searches for `pat', and filters current results.
Also narrow the pager view, and leave room for one head line (which is displayed
the prefixed line (if any)), and one for messages (the last one on screen).

Use this msg line to display the entered pattern. Note that by default we set the
IGNORE_CASE flag.

Also do not trigger automatically auto-completion on the first keypress, if the
key is in the a...z range. We are already too aggressive in the interaction.