Add an X development namespace.
This is quite early for that, as we make too many of assumptions, but we're walking
fast so to set the end intented border, (and if we're inclined to continue) and we
can then walk backwards to the first bit (it is called avroh and aroh i think).

Commited also some trivial changes gathered over the time.