packman: redirect build output to a logfile, but also print any errors to standard error stream.
Add a independent process() function to the sh-module, though we don't use the sh instance
there. This function was born because we wanted the above functionality, but the two
libraries (sh/proc) are too complicated to adjust and quite possible some code
would break because of those invasive changes.

Many more fixes and enhancements to the package manager. Because of them the code
is in an unstable stage.

Also libcurl and wolfssl seems that they don't cooperate well with eachother, with
various versions combinations. It might be because of us though.

This commit is incomplete, but we had to push it because of the local situationism.