Add array_prepend () to std-module. This has been implemented for String and Map Types only.
Language: fix a memory leak when a string literal is the first argument to an add operation.

Also fix? a memory leak when a map member of a map is an argument to a function. Release it
even if the refcount is < 0.

Also in the append() builtin, get the key with an expression, rather from the tokenstring.

Packman lib: add get_deps() and get_deps_rec() to get the dependencies of a package.

Enhance the Make bootstrap unit by copying necessary included files from Make. This is
a prerequisite when calling the script for the very first time it runs.

Fix the case when passing NULL to vsnprintf libc unit.