This is a proof of concept of a lexical parser, that parses current language units.
It can return the tokens in json format, which it is also valid code for the current
interpreter. It can also recreate the code unit argument, from the produced tokens
and can check for equality. A few tests were succesfull. Obviously there might be
corner cases or not so corner.

Obviously it can has many usages like to early check for syntax errors, or
to produse a more readable unit for the current interpreter (some complicated
code chunks (if/else or chained commands, putting braces, ...)), or even as a
way to speed up execution with a predone lexical analysis unit. And many more.

But it won't go this way.

Also avoid to build zsu executable by default. Unfortunatelly the docs they
didn't get this important update.

Also in libcontain, i've noticed that the CLONE_NEWTIME capability fails with
kernel < 6 and recent glibc.