libc: close the file descriptor in readfile_u() after completion.
Make: add a command line option to set a definition flag that can be used during
compilation (can be repeated as needed and it doesn't requires the -D prefix (it
is appended by Make)).

libe: do not highlight numbers that included in a keyword (that was pretty embarrassing).
However doesn't handle the case of a keyword that starts with a number (this almost never
happens in practice but it might be happen), just for performance reasons (requires a tmp
string to hold the parsing until to make sure it is really a number (and not a mix), and
only then it can be flushed (surrounded with a proper color for the occasion)).