This is to fix a previous mistake in one of the libc functions.
To prevent those kind of mistakes in future, now make outputs a C file
(as libc/libc/libc/libc.c) that includes all the libc functions.

This is just for testing reasons and it is not intented to be used as a
monolithic C library. It can be used, but we should expose the exported
functions in a header first.

A simple .build file that can be used to test the generated libc, might look
like:

var m = New Make (__file__, SHARED_LIBRARY|STATIC_LIBRARY)
m.set.library.source ("libc.c"; clear)
m.set.liboutput ("C")
m.set.header.source (null)
return m

At least for this commit both targets are compiled without a warning, with
the following sizes in x86_64:
530328 libC.a
275544 libC.so