Mostly trivial changes (though lots of them).
Add str_hash() variations and use the short version as the string hash function in map types.

Also add endian.h, dup(), flock() but without real testing.

Remove from libc all the special syntax sugar macros, with the `ifnot' exception.

Bump version to 0.4.4, which it is merely signals (the long time awaited of) the end
of the development as we know it.

Either we'll continue to another project and use this codebase as our development tool,
or we have to wait a bit to adjust in a proper state (though we are not that far
away from that, but we've to set up another scratch point and at the same time
to support the past (hard to do that though when you are riding on the wilderness
of the ocean drive all the way up to the red desert hills))). So we may want to stick
at this point and always use git --amend.