Extend vformat().
Add support for '#' alternate form which for 'b' prepends '0b',
for 'o' prepends '0' and for 'x' prepends '0x'. A 'p' directive
is always in this form.

Add support for '%.*s' which for strings means, that the next
int argument will be the number of bytes to write from the next
char * argument.

Add support for '%[n (int but not 0)]s' to mean that the next
char * argument will be padded with spaces to the left side, if
is smaller than the given 'n' field width.

Add support for '%.[n (int)]s' to mean that if the next char *
argument is bigger than the 'n' specifier, at most 'n' bytes will
be written.