This is to add, and after a long long time, a new editor functionality.
I had to touch this code for maybe years and thankfully was rather easy
to find the way through the api and the code structure and logic (it is
a bit strange to look up for your own (primitive at the time) code).

By the way a small unsolved puzzle:
The first comes in last. The last and the first will be first again.

Also commited (an almost finished prototype) a primitive memory allocator
for recording reasons (it seems we'll be in need for memory as we go old).