Add make libc[-force] targets to the top Makefile to generate the C library.
Expand the syscall macros for readability. Macros are hiding details, and
while for some cases like handling errors from function returns or in any
case there is repetitive same task for a common set of functions might be
indeed useful, in C details matters much, especially when the code is in
initial development. Finally it seems to be in C that is "hide as less is
possible" even for your own code (and which by default adds an even small
layer of self confidence), for your own good (if it happens to face again
your own code). But this is C and for all function calls you need to know
their details, not only the signature but and inner details. That's C and
that is that you love in C, and that is why the code structure matters. I
really appreciate boring C when I see C code.