Make the new allocator the default to test its functionality.
We first tried this allocator to build the   static version of our
language, which is exact code with the one we use currently except
of dynamic loading (which anyway is disabled in static builds), and
which totally depends on our libc and our memory allocator.

But first, this new allocator is using 24 bytes less than the old,
though because it can not handle more than UINT_MAX, but still we
have an unsigned int and a short int free to use in the structure.

The first working unoptimized prototype roughly performed a little
(10%) better than the old one. However soon it was realized that the
actual cost was the lookup function, because more and more chunks at
the beginning (and with time) were all occupied by usually constants
or initialization kind of allocations.

So we simply put some code to always store the first unused chunk, so
the next lookup would always start from that free chunk. We applied
this almost free of cost optimization and we got many times faster
results, which match that of the static language version which uses
system's libc, and this is some satisfaction as it balances somehow
with the dissapointment we faced when we show the first results when
we first builded this own libc version of the language.

But as we gradually programming this allocator, this time we observed
the behavior and we then adapt, we didn't the same with the old code.