Fixing a couple of bugs that revealed by the new allocator.
Almost all of them it was because when were allocated pointers, we didn't
initialized them to zero, as this was done by default from the allocator
(by both the old and the new one at the beginning).

We fixed also the mem_set() function and replaced it with mem_set_all()
which works like standard memset. We've added an explicitly Calloc, but
which unlike standard calloc, accepts only one argument (the new size).

The thing is that, when you depend in a function behavior, sometimes you
forget your C duties, so it is desirable (in a way) to treat your C as:
what the code will do, is what you see. Explicitly matters a lot.

Anyway the ship has been sailed and may to find the strength to stand up
and fight for a common cause. We've to find this cause and cooperate, or
else I'm afraid will be too late to fix the damages as the time matters.