Add talloc.
This is a Tiny allocator which is like the zalloc.[ch] implementation
but with one actual difference.

Both the header size and the alignment is at 8 bytes, so the maximum
allocated size is at UINT_MAX minus 7 and the minimum chunk size that
is allocated is at 16 bytes.

However in this implementation (which has no specific field to store
the un|used flag) we use the unused lowest order bit of the chunk_size
struct member (because of the alignment this bit is always zero), and
so the calculations are different. The semantics however should be the
same and so and the underlying algorithm.

Running the tests of the language it uses 48K less than the 16 bytes
version and performs the same (though it hadn't to be that way, as it
uses extented calculations).