Add falloc allocator.
This was intented to be a Fast allocator because it uses a specialized
field to set the [un]used flag and two pointers to locate the next and
previous chunks (which results to 24 bytes memory overhead), but wasn't
the case. It performs at its best equal to talloc which uses arithmetic
to find those chunks and bit manipulation to store the flags, while it
uses 8 bytes.

As it was tricky to get this thing right, it stays as a reference in the
repository for future perhaps usage.

By testing the implementations, it was revealed that raising the extra bytes
at 4096 instead of 1024 (when we are allocating new chunks from the kernel),
all they were able to perform two bits better, so calls to sbrk() cost.