Test blocks can now be inlined into the libc ".c" units.
Add dirwalk.[ch] interface and be the first to test the functionality.

For now tests can be embedded in units like:
/* test {
   ... test code
} */

Test blocks can be anywhere but below the unit headers.

A file with the same name is written under the tests directory.

For now tests can be run with:

  La TestLibc.lai

from the tests directory (but this should change at some point).

What it could be useful, is to use differently a unit with a main()
or without (ideally those units should be treated with both ways).