Add file/copy(), file/is_link(), file/is_readable(), file/size(), std/fchmodat(), std/fchownat(), std/link(), std/mknod(), std/symlink(), std/utimensat(), utf8_change_case(), utf8_get_code_at(), fd_write (), fd_read(), file_is_link().
When generating libc, try to get the macros right from units as they wish
to declare them, by looking to their headers, instead of generating them
by the unit name. Also implement the skip and as tokens for header units.

Sync the syscall numbers from asm/unistd_[32|64].h

dirlist() now is a wrapper for dirwalk() with depth=1.

Add an integer array to the dirwalk() to store the types and also always
canonicalize the file_name with realpath().

L:
  - the integer type now is int64_t wide
  - remove the useless Types from modules, instead add a variable to the
    std function to fool the import function
   - implement almost all of the static std-module and add the io-module