vformat() development.
This already went too far (it is not such tiny enymore), however most
of the features are good to have. But however, we might included some
bugs, as the unit is exactly or close at the point that can easily be
characterised as a bit more complicated than perhaps it should.

And while we've tried to catch the obvious errors (perhaps some bits more
than the obvious ones), but still there are corner cases or not so corner
present in the code, and with this commit we might introduced some more.

However in C, variadic functions (especially in the format family) are an
easy target for disaster (mostly because it's rather easy to get the number
of arguments wrong or in a wrong sequence), as we don't use only them just
for printing, but also to construct accurate strings for serious business.

And just for this i could write some thousand lines of code to parse C code
just to get them right. I'm not that afraid of out of bounds operations or
to miss a null byte at the end of the string or off by one errors, as when
using such functions. Interpolated strings are pleasant to work with, and
with much much much smaller probabity to get then wrong, as the arguments
are inlined in the string, and they don't go far away to the right corner
or to the next line. In any case there isn't a gap of your intention and
and the argument, but how they call it? Locality.