Add the base64_* functions (the same that we already use in z namespace) to libc.
Add also a bit more than a couple functions to libc.

Add gdb support to Libc.test.

Install any data of an executable also for the static targets with the suffix -static.
We probably want to turn them around and we use the static executables as the default
symbolic links to commands.

Freestanding static executables from now and on it is our only consideration. Reasons:

  - we do not use libraries (only libc)
  - the size overhead is ridiculous lower (our static freestanding language weighs
    309960 vs 2527480 bytes)
  - simplifies everything (except dependencies to functions that change, we could do
    probably something for this if time is ridiculous generous)
  - the nature of our environment doesn't benefit much from the advantages of shared
    libraries,

But most. Our project was born and live for the realization of the mechanics.

For instance the recent addition of CRT (c runtime environment), shows which really
is the entry point function.in C. We could take advantages from this of many kinds.

C is also a perfect also example as C by itself is a standard in software universe,
plus is low level enough to understand the mechanics of the hardware, with a plain
hermetic syntax, enough though to be pleasant to write and read. But is not easy.
One of the main things you have to understand is the underlying types.
Some of them unfortunate established with a wrong way.

Take char:
   - ambiguous name as of utf8 days. char (for character i assume) is not represented
     always by one byte (can be up to four). One byte (8 bits) is the width of the type

   - char is not standardized to be either signed or unsigned, as it depends of the
     machine. The machines convention is char to be signed.

     We have this in std/int.h

     #define SCHAR_MIN     (-128)
     #define SCHAR_MAX     127
     #define UCHAR_MAX     255

     #ifdef __CHAR_UNSIGNED__
     #define CHAR_MIN      0
     #define CHAR_MAX UCHAR_MAX
     #else
     #define CHAR_MIN SCHAR_MIN
     #define CHAR_MAX SCHAR_MAX
     #endif

     So that means if CHAR_MIN is 0 then char in this case is unsigned.

     We can force gcc with "-funsigned-char" or "-fsigned-char" to handle char as
     either unsigned or signed respectively.

     So in the clause above that we use to define its range, we could also do:

     #ifdef __CHAR_UNSIGNED__
       typedef unsigned char  byte;
     #else
       typedef signed   char  byte;
     #endif
And that we did and it looks fair, because (char *) it really points to a byte.

We also added  (a bit overexaggerated maybe)
static_assert (sizeof (char)      == 1, "error: awaiting size of a char to be 1 bytes long");
static_assert (sizeof (short)     == 2, "error: awaiting size of a short to be 2 bytes long");
static_assert (sizeof (int)       == 4, "error: awaiting size of an int to be 4 bytes long");

static_assert (sizeof (long)      == 8, "error: awaiting size of a long to be 8 bytes long");
static_assert (sizeof (size_t)    == 8, "error: awaiting size of a size_t to o be 8 bytes long");
static_assert (sizeof (size_t)    == 4, "error: awaiting size of a size_t to o be 4 bytes long");
static_assert (sizeof (long)      == 4, "error: awaiting size of a long to be 4 bytes long");
That is to be sure for the expectations, else fail hard at compile time.

Static_assert is one of the main reasons we compile in c11 and onwards. C99 doesn't
have it. It can be emulated a bit from a quick glance in internet but solutions are
complicated. Honestly the better C is C11 plus C23's nullptr and bool (both of them
they should be defined from the very begining). nullptr probably as null though (we
know that is a pointer).

Because of these and other reasons we introduced:

typedef uint8_t   u8;
typedef uint16_t  u16;
typedef uint32_t  u32;
typedef uint64_t  u64;
typedef size_t    usize;

typedef int8_t    i8;
typedef int16_t   i16;
typedef int32_t   i32;
typedef int64_t   i64;
typedef ssize_t   isize;

typedef float     f32;
typedef double    f64;

It seems better to use them, as they match the number of bits, in a consistent
established way.
And since that virgil provides integer types for all bit widths from 1 ... 64

     https://github.com/titzer/virgil

then without doubt we should better use them.

Commited also a couple of fixes.